NSS: Introduce gethostbyname5_r function
Add af (address family) to the list of arguments of gethostbyname4_r
function in files, dns, nis and nisplus module and rename the function to
gethostbyname5_r. Pass the af value to the underlying parsing functions.
Define gethostbyname4_r as a wrapper for the former passing af set
to AF_UNSPEC.

Signed-off-by: Łukasz Stelmach <l.stelmach@samsung.com>